refactor: Внесёт правки по замечаням наставника
