+++ project/apps/users/docker-compose.dev.yml
@@ -1,7 +1,7 @@
 version: '3.5'
 services:
   taskforce.users.mongo:
-    image: mongo
+    image: mongo:4.2
     restart: always
     container_name: taskforce.users.mongo
