+++ project/apps/users/src/app/app.module.ts
@@ -1,10 +1,9 @@
 import { Module } from '@nestjs/common';
 import { TasksUserModule } from './tasks-user/tasks-user.module';
 import { AuthenticationModule } from './authentication/authentication.module';
-import { ConfigUsersModule } from '@project/config/config-users';
+import { ConfigUsersModule, getMongooseOptions } from '@project/config/config-users';
 import { DevtoolsModule } from '@nestjs/devtools-integration';
 import { MongooseModule } from '@nestjs/mongoose';
-import { getMongooseOptions } from '../../../../libs/config/config-users/src/lib/get-mongoose-options';
 
 @Module({
   imports: [
