+++ project/apps/users/src/app/authentication/authentication-controller-messages.ts
@@ -1,0 +1,6 @@
+export enum AuthenticationControllerMessages {
+  USER_CREATED = 'The new user has been successfully created',
+  USER_LOGGED_IN = 'User has been successfully logged.',
+  WRONG_PASSWORD_OR_LOGIN = 'Password or Login is wrong.',
+  USER_FOUND = 'User found',
+}
