+++ project/libs/config/config-users/src/index.ts
@@ -1,3 +1,4 @@
 export { default as dbConfig } from './lib/config/db.config';
 export { default as appConfig } from './lib/config/app.config';
-export * from './lib/config-config-users.module';
+export * from './lib/config-users.module';
+export * from './lib/config/get-mongoose-options'
