+++ project/libs/config/config-users/src/lib/config/get-mongoose-options.ts
